Second step autocomplite fix
